Formatting style.

+++ source/2023/09/solver-2023-09.cpp
@@ -1,10 +1,8 @@
 #include <aoc.hpp>
 
 #include <absl/container/btree_map.h>
-#include <absl/container/flat_hash_map.h>
 #include <absl/container/flat_hash_set.h>
 #include <fmt/format.h>
-#include <fmt/ostream.h>
 #include <re2/re2.h>
 
 namespace {
@@ -42,7 +40,7 @@ auto advent<2023, 9>::solve() -> Result {
 
     // Part 1 & 2
     sequences.back().push_back(0);
-    for (i64 i = sequences.size() - 2; i >= 0; i--) {
+    for (i64 i = static_cast<i64>(sequences.size() - 2); i >= 0; i--) {
       sequences[i].emplace_back(sequences[i].back() + sequences[i + 1].back());
       sequences[i].emplace_front(sequences[i].front() - sequences[i + 1].front());
     }
